docs: dynanic load components to improve first screen render
