+++ lib/components/layout/powered-by/index.ts
@@ -1,0 +1,5 @@
+import dynamic from 'next/dynamic'
+
+const PoweredBy = dynamic(import('./powered-by'), { ssr: false, loading: () => null })
+
+export default PoweredBy
