+++ next.config.js
@@ -5,6 +5,10 @@ const withMDX = require('@next/mdx')({
   },
 })
 
+const withBundleAnalyzer = require('@next/bundle-analyzer')({
+  enabled: process.env.ANALYZE === 'true',
+})
+
 const nextConfig = {
   generateEtags: false,
 
@@ -65,4 +69,4 @@ const nextConfig = {
   },
 }
 
-module.exports = withMDX(nextConfig)
+module.exports = withBundleAnalyzer(withMDX(nextConfig))
